ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
mysql-connector-python==8.2.0
python-dateutil==2.8.2
joblib==1.3.2